implemented logging with log4j2, arg4j for handling arguments and options and spock for testing. partially fixed p-value bug in parallel execution.

+++ src/main/java/org/mpii/jami/CMIComplete.java
@@ -117,7 +117,7 @@ public class CMIComplete{
         RecursiveTask<Integer> task = new CMIRecursiveTask(numPerm, cmi, ip, origData.get(2).length);
         int sum = commonPool.invoke(task);
         sum=Math.max(sum, 1);
-        pValue=sum/numPerm;
+        pValue=(double) sum/numPerm;
     }
 
     public void computeCMIandPValueBetterPartitioning(){
